rrForcedDepth introduced
Separate the point at which russian roulette termination (based on the path
throughput) is started from the point at which a non-zero termination
probability is forced (and allow for this forcing to never happen if
desired).

This gives better behaviour for scenes that require long paths (e.g.
multiple scattering in a medium) but which are guaranteed to have a
decrease in throughput (e.g. no total internal reflection cycles possible,
so that the external forcing of the path termination is not neccesary to
keep the tracing in finite time).