Added rrTargetThroughput + RussianRoulette abstraction
The russian roulette path termination criterion is augmented with a
'rrTargetThroughput' option, wich describes the throughput level where RR
kicks in (it tries to keep the throughput at or above that level). This is
useful for scenes where all visible contributions come from long paths with
low throughput.

The russian roulette mechanism is abstracted away in a RussianRoulette
structure to avoid code clutter/duplication as much as possible.